requirements.txt
selenium>=4.20.0
webdriver-manager>=4.0.1
google-generativeai>=0.8.0
python-dotenv>=1.0.1
